datastore package refactor
Moved the firebase/firestore package to under datastore as it represents
an implementation of datastore.
