+++ auth/user/firebase_manager.go
@@ -19,7 +19,7 @@ import (
 
 	"cloud.google.com/go/firestore"
 	"github.com/qqiao/webapp/datastore"
-	f "github.com/qqiao/webapp/firebase/firestore"
+	f "github.com/qqiao/webapp/datastore/firestore"
 	"google.golang.org/api/iterator"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
